mutex: Implement `Mutex` on types, not references
The previous implementation implemented the `mutex_trait::Mutex` trait
on `&T`, where `T` is any of the available mutex lock types. Since
`mutex_trait::Mutex::lock` needs a `&mut self` to work on, this makes
usage of `Mutex` in functions that accept a mutex parameter impossible:
When the mutex is passed in as `&T` (To satisfy the trait
implementation), `lock()` cannot be called since `&T` cannot be borrowed
as `&mut T`. On the other hand, when passed in as `&mut T`, it cannot
use the `Mutex` impl, since that is only defined on `&T`.

Instead, now we implement the Mutex trait on `T`, as a value. This makes
passing Mutexes to functions possible.